+++ mobile/alix/rust/src/frb_generated.rs
@@ -1571,6 +1571,9 @@ const _: fn() = || {
         let _: u32 = ReviewState.passed;
         let _: u32 = ReviewState.failed;
         let _: u32 = ReviewState.acquired;
+        let _: u32 = ReviewState.recognized;
+        let _: u32 = ReviewState.recognize_partly;
+        let _: u32 = ReviewState.recognize_missed;
         let _: bool = ReviewState.can_restart;
         let _: bool = ReviewState.promotable;
         let _: Option<u64> = ReviewState.next_due_ms;
@@ -2349,6 +2352,9 @@ impl SseDecode for crate::api::review::ReviewState {
         let mut var_passed = <u32>::sse_decode(deserializer);
         let mut var_failed = <u32>::sse_decode(deserializer);
         let mut var_acquired = <u32>::sse_decode(deserializer);
+        let mut var_recognized = <u32>::sse_decode(deserializer);
+        let mut var_recognizePartly = <u32>::sse_decode(deserializer);
+        let mut var_recognizeMissed = <u32>::sse_decode(deserializer);
         let mut var_canRestart = <bool>::sse_decode(deserializer);
         let mut var_promotable = <bool>::sse_decode(deserializer);
         let mut var_nextDueMs = <Option<u64>>::sse_decode(deserializer);
@@ -2374,6 +2380,9 @@ impl SseDecode for crate::api::review::ReviewState {
             passed: var_passed,
             failed: var_failed,
             acquired: var_acquired,
+            recognized: var_recognized,
+            recognize_partly: var_recognizePartly,
+            recognize_missed: var_recognizeMissed,
             can_restart: var_canRestart,
             promotable: var_promotable,
             next_due_ms: var_nextDueMs,
@@ -3110,6 +3119,9 @@ impl flutter_rust_bridge::IntoDart for FrbWrapper<crate::api::review::ReviewStat
             self.0.passed.into_into_dart().into_dart(),
             self.0.failed.into_into_dart().into_dart(),
             self.0.acquired.into_into_dart().into_dart(),
+            self.0.recognized.into_into_dart().into_dart(),
+            self.0.recognize_partly.into_into_dart().into_dart(),
+            self.0.recognize_missed.into_into_dart().into_dart(),
             self.0.can_restart.into_into_dart().into_dart(),
             self.0.promotable.into_into_dart().into_dart(),
             self.0.next_due_ms.into_into_dart().into_dart(),
@@ -3933,6 +3945,9 @@ impl SseEncode for crate::api::review::ReviewState {
         <u32>::sse_encode(self.passed, serializer);
         <u32>::sse_encode(self.failed, serializer);
         <u32>::sse_encode(self.acquired, serializer);
+        <u32>::sse_encode(self.recognized, serializer);
+        <u32>::sse_encode(self.recognize_partly, serializer);
+        <u32>::sse_encode(self.recognize_missed, serializer);
         <bool>::sse_encode(self.can_restart, serializer);
         <bool>::sse_encode(self.promotable, serializer);
         <Option<u64>>::sse_encode(self.next_due_ms, serializer);